Function build_const::write_array_raw

pub fn write_array_raw<W: Write>(
    w: &mut W, 
    name: &str, 
    ty: &str, 
    raw_values: &[&str]
) -> String

Write an array of raw values and return the array's full type representation.

This can be used to create nested array constant types.